Undefined: locator_source in MapContributionCapture
This was commented out by me in d03c199fdcf794ea4f119ebedd4feff19d526c0e
because it threw errors in development and seemed to work fine without it.
(country is in search results)
But: it meant that locator_source is not there for the click handler.
(and no eslint error to alert me of the problem).

Now it does set locator_source, catches and handles possible errors.